Change precipitation_probability data
Using the current precipitation probability is pretty useless.  I moved it to the probability for the day.  It might be nice to have both a daily and hourly or just restructure this whole sensor so that you can add current, minutely, hourly for each.  It's more flexible and less code that way.